Type agnostic, finite recursive_bottom_eltype function
Before the `recursive_bottom_eltype` function was written with only
numeric types in mind. For those, it worked. But for non-numeric types
it recused infinitely, causing a `StackOverflowError`.
Now it works on any type and is sure to recurse only a finite number of
time, thus always end and return a result.

I also added tests that cover this specific function, testing numbers
with and without units as well as chars and strings.