Fix StructArray broadcast in VectorOfArray
Fixes https://github.com/SciML/RecursiveArrayTools.jl/issues/410

This specializes so that if `u.u` is not a vector, it will convert the broadcast to fix that. I couldn't find a nice generic way to use `map` so the fallback is to build the vector and convert, which seems to not be a big performance issue. For StructArrays, `convert(typeof(x), Vector(x))` fails, and so this case is specialized.