More precise deprecations of linear indexing
Noticed in https://github.com/SciML/Catalyst.jl/pull/1248, the linear indexing is actually fine when done on scalar arrays because that matches the linear indexing that you get from an AbstractArray interface. So instead the deprecations are made more precise to be just about the array of array case.